requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autoflake==2.3.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.0
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.4
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-multiselectfield==0.1.13
django-parler==2.3
django-parler-rest==2.2
django-ranged-response==0.2.0
django-redis==5.4.0
django-rosetta==0.10.1
django-simple-captcha==0.6.0
django-taggit==6.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
django-safedelete==1.4.1
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.28.0
executing==2.1.0
Faker==35.2.0
fonttools==4.55.3
html5lib==1.1
idna==3.10
inflection==0.5.1
ipython==8.31.0
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
polib==1.2.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyphen==0.17.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
ruff==0.8.6
sentry-sdk==2.19.2
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stack-data==0.6.3
stripe==11.4.1
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
gunicorn==23.0.0
flower==2.0.1
